# Fix test explorer performance issues