fixes https://github.com/scalameta/metals/issues/3501

Previously, due to lack of caching and https://github.com/scalameta/metals/issues/3509 discovery request were able to take over 2 seconds in Metals repo.

Now, only the first request takes a significant amount of time (due to https://github.com/scalameta/metals/issues/3509 additional lookups for non-existing symbols may happen) and every subsequent request takes constant, close to zero time.

```shell
2022.01.14 14:21:18 INFO  Cache refresh took: 1752ms
2022.01.14 14:21:25 INFO  Cache refresh took: 1ms
2022.01.14 14:21:25 INFO  Cache refresh took: 1ms
```

From what I observed this [code](https://github.com/scalameta/metals/blob/5898e3b5bd4d56a4e930e37c5cb3fba8366af055/metals/src/main/scala/scala/meta/internal/metals/Compilations.scala#L220-L223) was calling refresh a few times within less than a second at the start of Metals. In order to avoid problems with unnecessary computing, I used a single-thread executor. After first run which populates caches, subsequent ones don't have much work. The executor also ensures that requests from the client will always be responded with the newest data. Maybe someone has some ideas on how to improve it?

There's one corner case that comes into my mind - if someone will rename a file without renaming the test suite then the wrong location will be cached for that test suite. I'll try to do something in this scenario.